alert added to feedback.js

+++ src/js/03-feedback.js
@@ -13,19 +13,27 @@ const checkStorage = () => {
 };
 checkStorage();
 
+const validateEmail = email => {
+  const regex = /\S+@\S+\.\S+/;
+  return regex.test(email);
+};
+
 const submitForm = e => {
   e.preventDefault();
   const obj = {
     email: email.value,
     message: message.value,
   };
+  if (validateEmail(email.value) === true) {
+    localStorage.setItem('feedback-form-state', JSON.stringify(obj));
+    console.log(obj);
 
-  localStorage.setItem('feedback-form-state', JSON.stringify(obj));
-  console.log(obj);
-
-  localStorage.clear();
-  email.value = '';
-  message.value = '';
+    localStorage.clear();
+    email.value = '';
+    message.value = '';
+  } else {
+    alert('Please type in correct email address: example@go.com');
+  }
 };
 
 email.addEventListener(
